Generalise GBRT based optimisation to all tree based models
